💄 Simplify the share dialog to an email row, invitees and a link footer
