+++ apps/web/public/locales/en/app.json
@@ -402,7 +402,6 @@
   "invalidSessionTitle": "Your session is no longer valid",
   "inviteCanceledSuccess": "Invite canceled successfully",
   "inviteFailed": "Failed to send invitation",
-  "inviteLink": "Invite link",
   "inviteLinkCopied": "Invite link copied",
   "inviteMember": "Invite member",
   "inviteMemberDescription": "Invite a new member to your space",
@@ -414,7 +413,6 @@
   "inviteMemberFormSubmit": "Send invite",
   "inviteMembersUpsellDescription": "Invite members to manage polls and events together in this space.",
   "inviteNotEnoughSeats": "There are not enough seats available to send this invite",
-  "inviteParticipantLinkInfo": "Anyone with this link will be able to vote on your poll.",
   "inviteParticipantsDescription": "Copy and share the invite link to start gathering responses from your participants.",
   "invitePending": "An invitation has already been sent to this email address",
   "inviteSent": "Invitation sent",
@@ -778,11 +776,11 @@
   "share": "Share",
   "shareDialogAlreadyInvited": "{email} is already invited",
   "shareDialogAlreadyResponded": "{email} has already responded",
+  "shareDialogAnyoneWithLink": "Anyone with the link",
+  "shareDialogCanRespond": "Can respond",
   "shareDialogDailyLimit": "You've reached today's limit of 100 invites",
   "shareDialogGuestPrompt": "Create an account to invite people by email. Your invites and their responses stay attached to your account instead of this browser.",
   "shareDialogInvalidEmail": "Enter a single valid email address",
-  "shareDialogInviteByEmail": "Invite by email",
-  "shareDialogInviteByLink": "Invite by link",
   "shareDialogInvitedCount": "{count, plural, one {# invited} other {# invited}}",
   "shareDialogInviteRevokedAnnouncement": "Invite for {email} revoked. {count}",
   "shareDialogInviteSentAnnouncement": "Invite sent to {email}. {count}",
